pyproject.toml
[project]
name = "rd-station"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Hernandes",email = "gabriel.hernandess01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "django-restframework (>=0.0.1,<0.0.2)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
